vet ikke

+++ Visualize/Article2/ROC/plot_roc_curve_matrix_G1.m
@@ -1,5 +1,4 @@
-function h_fig = plot_roc_curve_matrix_G1(ROC,...
-		classifiers_to_use, titleStr)
+function h_fig = plot_roc_curve_matrix_G1(ROC, classifiersToUse, titleStr)
 	% Specific/max inflations by speeds in subplots
 	% Plot ROC curves for states of specific or pooled intervention levels.
 	%
@@ -10,21 +9,19 @@ function h_fig = plot_roc_curve_matrix_G1(ROC,...
 	% Curves overlaid in each panels for:
 	% - NHA as classifier, component-wise
 
-	classifier_labels = classifiers_to_use(:,2);
-	classifier_inds = find(ismember(ROC.classifiers,classifiers_to_use(:,1)));
+	classifierLabels = classifiersToUse(:,2);
+	classifier_inds = find(ismember(ROC.classifiers,classifiersToUse(:,1)));
 	classifiers = ROC.classifiers(classifier_inds);
 	
 	spec = get_plot_specs;
-	figHeight = 1400;%900; %955
-	figWidth = 1180;
-	panelLength = 154;
-	gap = 25;
-	
-	% Offsets for common axes
-	mainAxGap = 15;
-	
+		
 	nRows = size(ROC.predStates,1);
-	nCols = numel(ROC.speeds);
+	nCols = size(ROC.X,2);
+	figHeight = 1050;%900; %955
+	figWidth = (750/3)*nCols+430;
+	panelLength = 150;
+	gap = 20;
+	mainAxGap = 15;
 	h_fig = figure(spec.fig{:},...
 		'Name',titleStr,...
 		'Position',[10,40,figWidth,figHeight]);
@@ -43,50 +40,25 @@ function h_fig = plot_roc_curve_matrix_G1(ROC,...
 				Y = ROC.Y{i,j,classifier_inds(k)};
 				AUC = ROC.AUC{i,j,classifier_inds(k)};
 
-% 				h_ax(i,j).ColorOrderIndex = k;
-% 				h_nha(k) = plot(h_ax(i,j),X(:,1),Y(:,1));
+				h_ax(i,j).ColorOrderIndex = k;
+				h_nha(k) = plot(h_ax(i,j),X(:,1),Y(:,1));
 
-				if contains(classifiers{k},'speed')
-					h_ax(i,j).ColorOrderIndex = 1;
-					h_nha(k) = plot(h_ax(i,j),X(:,1),Y(:,1),'Color',[.5 .5 .5]);
-					set(h_nha(k),'LineWidth',5);
-   					%plot_shaded(X(:,1),Y(:,1),'LineWidth',0.1,'Color',h_nha(k).Color);
-  					%text(0.625,0.3,[num2str(AUC(1),'%1.2f')],spec.rocText{:})
-				elseif contains(classifiers{k},'norm')
-					h_ax(i,j).ColorOrderIndex = k;
-					h_nha(k) = plot(h_ax(i,j),X(:,1),Y(:,1));
-					set(h_nha(k),'LineWidth',5);
-%    					area(X(:,1),Y(:,1),'FaceColor',h_nha(k).Color,...
-%    						'EdgeColor','none', 'FaceAlpha',0.06)
-% 					plot_shaded(X(:,1),Y(:,1),'LineWidth',0.1,'Color',h_nha(k).Color);
-%  					text(0.625,0.3,[num2str(AUC(1),'%1.2f')],spec.rocText{:})
-				elseif contains(classifiers{k},'sumNHA')
-					h_ax(i,j).ColorOrderIndex = k;
-					h_nha(k) = plot(h_ax(i,j),X(:,1),Y(:,1));
-					set(h_nha(k),'LineWidth',5);
+				if contains(classifiers{k},'sum')
+					set(h_nha(k),'LineWidth',4);
    					area(X(:,1),Y(:,1),'FaceColor',h_nha(k).Color,...
-   						'EdgeColor','none', 'FaceAlpha',0.06)
+						'EdgeColor','none', 'FaceAlpha',0.06,'HandleVisibility','Off')
 % 					plot_shaded(X(:,1),Y(:,1),'LineWidth',0.1,'Color',h_nha(k).Color);
   					text(0.625,0.3,[num2str(AUC(1),'%1.2f')],spec.rocText{:})
-				elseif contains(classifiers{k},'x')
-					h_ax(i,j).ColorOrderIndex = 2;
-					h_nha(k) = plot(h_ax(i,j),X(:,1),Y(:,1));
-					set(h_nha(k),'LineWidth',2,'LineStyle',':','Color',min(1.1*h_nha(k).Color,[1 1 1]));
-				elseif contains(classifiers{k},'y')
-					h_ax(i,j).ColorOrderIndex = 3;
-					h_nha(k) = plot(h_ax(i,j),X(:,1),Y(:,1));
-					set(h_nha(k),'LineWidth',2,'LineStyle',':','Color',min(1.1*h_nha(k).Color,[1 1 1]));
-				elseif contains(classifiers{k},'z')
-					h_ax(i,j).ColorOrderIndex = 4;
-					h_nha(k) = plot(h_ax(i,j),X(:,1),Y(:,1));
-					set(h_nha(k),'LineWidth',2,'LineStyle',':','Color',min(1.1*h_nha(k).Color,[1 1 1]));
+				elseif contains(classifiers{k},'norm')
+					set(h_nha(k),'LineWidth',2.25,'LineStyle',':','Color',min(1.1*h_nha(k).Color,[1 1 1]));
+				elseif contains(classifiers{k},'best')
+					set(h_nha(k),'LineWidth',2.25,'LineStyle',':','Color',min(1.1*h_nha(k).Color,[1 1 1]));
 				elseif contains(classifiers{k},'P_LVAD')
-					h_ax(i,j).ColorOrderIndex = 5;
-					h_nha(k) = plot(h_ax(i,j),X(:,1),Y(:,1));
 					if i<nRows, set(h_nha(k),'Visible','on'); end
-					set(h_nha(k),'LineWidth',2,'LineStyle',':',...
+					set(h_nha(k),'LineWidth',1.5,'LineStyle',':',...
 						'Color',[h_nha(k).Color,0.9]);
 				end
+
 
 			end
 
@@ -108,22 +80,28 @@ function h_fig = plot_roc_curve_matrix_G1(ROC,...
 	set(h_ax,spec.rocAx{:})
 	h_ax = position_panels(panelLength,h_ax,gap);
 	offset_main_ax(h_ax,h_xax,h_yax,mainAxGap,mainAxGap);			
-	format_tick_labels(h_ax,h_xax, h_yax);
+  	format_tick_labels(h_ax, h_xax, h_yax);
+	set(h_ax,'YTickLabel',{},'XTickLabel',{},'GridLineStyle','-','GridColor',[.9 .9 .9]);
+	set(h_ax,'LineWidth',1,'XTick',0:0.2:1,'YTick',0:0.2:1);
 	
-	h_leg = legend(h_nha, classifier_labels, spec.leg{:});
-	add_subtitles(h_ax,cellstr("\bf"+string(ROC.speeds)+"\rm RPM"),panelLength,gap,spec);
-	
+	h_leg = legend(h_ax(end), classifierLabels, spec.leg{:});
+	if ROC.pooledSpeed
+		add_subtitles(h_ax,"\bf{All}\rm RPM",panelLength,gap,spec);
+	else
+		add_subtitles(h_ax,cellstr("\bf"+string(ROC.speeds)+"\rm RPM"),panelLength,gap,spec);
+	end	
 	set([hXLab,hYLab],spec.supXLab{:})
-	hYLab.Position(1) = hYLab.Position(1)-20;
-	hXLab.Position(2) = hXLab.Position(2)+14;
+	hYLab.Position(1) = 60;
+    hXLab.Position(1) = 0;%(10/3)*nCols;
 	h_leg.Position(1) = h_leg.Position(1) + 275;
-	h_leg.Position(2) = 0;
-  		
+
+	h_leg.Position(2) = 0;%hXLab.Position(2);
+
 function h_ax = position_panels(panelLength,h_ax,gap_x,gap_y)
 	
 	if nargin==3, gap_y = gap_x; end
 
-	rowStart = 85;
+	rowStart = 80;
 	colStart = 80;
 	[nRows,nCols] = size(h_ax);
 	for i=1:nRows
